Complete functionality related to primitives
Primitive return types and parameter types are fully supported.
Object parameter types are supported.  However, Object return types
are not yet available.  Also, the only modifies to classes or methods
currently supported are public and final, though these are checked
(via the type system) as well as for Java source.