Partial rewrite of BC and new unit tests
The rewrite of the BC trait is with regards to how access modifiers
and class extends and implements clauses are handled.  Before, separate
objects were used, and in some cases created, to represent each group
of possible transitions.  Now, a single member object is used, where
different trait return types ensure invalid modifier and clause
combinations are not used.